+++ src/components/PasswordStrengthMeter.js
@@ -1,37 +1,37 @@
-import React from 'react'
-import zxcvbn from 'zxcvbn'
-import { withStyles } from '@material-ui/core/styles'
-import LinearProgress from '@material-ui/core/LinearProgress'
-import Typography from '@material-ui/core/Typography'
-import classNames from 'classnames'
+import React from 'react';
+import zxcvbn from 'zxcvbn';
+import { withStyles } from '@material-ui/core/styles';
+import LinearProgress from '@material-ui/core/LinearProgress';
+import Typography from '@material-ui/core/Typography';
+import classNames from 'classnames';
 
 const styles = (theme) => ({
   alert: {
     color: theme.palette.primary.main,
   },
-})
+});
 
 const PasswordStrengthMeter = (props) => {
-  const { password, classes } = props
+  const { password, classes } = props;
 
-  const pwdScore = zxcvbn(password).score
+  const pwdScore = zxcvbn(password).score;
 
   const pwdAlert = (pwdScore) => {
     switch (pwdScore) {
       case 0:
-        return 'Too weak'
+        return 'Too weak';
       case 1:
-        return 'Weak'
+        return 'Weak';
       case 2:
-        return 'Fair'
+        return 'Fair';
       case 3:
-        return 'Good'
+        return 'Good';
       case 4:
-        return 'Strong'
+        return 'Strong';
       default:
-        return 'Weak'
+        return 'Weak';
     }
-  }
+  };
 
   return (
     <>
@@ -42,11 +42,13 @@ const PasswordStrengthMeter = (props) => {
             color={classNames(pwdScore > 1 ? 'primary' : 'secondary')}
             value={pwdScore * 25}
           />
-          <Typography className={classes.alert}>{pwdAlert(pwdScore)}</Typography>
+          <Typography className={classes.alert}>
+            {pwdAlert(pwdScore)}
+          </Typography>
         </>
       )}
     </>
-  )
-}
+  );
+};
 
-export default withStyles(styles)(PasswordStrengthMeter)
+export default withStyles(styles)(PasswordStrengthMeter);