+++ src/components/SpeciesMgt.js
@@ -1,15 +1,19 @@
-import React from 'react'
-import SpeciesTable from './SpeciesTable'
-import { Grid } from '@material-ui/core'
+import React from 'react';
+import SpeciesTable from './SpeciesTable';
+import { Grid } from '@material-ui/core';
 
 const SpeciesMgt = () => {
   return (
-    <Grid container direction="column" style={{ flexWrap: 'nowrap', height: '100%' }}>
+    <Grid
+      container
+      direction="column"
+      style={{ flexWrap: 'nowrap', height: '100%' }}
+    >
       <Grid item container style={{ height: '100%', overflow: 'hidden' }}>
         <SpeciesTable />
       </Grid>
     </Grid>
-  )
-}
+  );
+};
 
-export default SpeciesMgt
+export default SpeciesMgt;