+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.1.2](https://github.com/Greenstand/treetracker-admin-client/compare/v1.1.1...v1.1.2) (2021-02-24)
+
+
+### Bug Fixes
+
+* console errors on login ([#19](https://github.com/Greenstand/treetracker-admin-client/issues/19)) ([b1322e1](https://github.com/Greenstand/treetracker-admin-client/commit/b1322e19df5f2d1ca25f0e8845d3539e5c1cbf15))
+
+## [1.1.1](https://github.com/Greenstand/treetracker-admin-client/compare/v1.1.0...v1.1.1) (2021-02-23)
+
+
+### Bug Fixes
+
+* planter org id filter ([#21](https://github.com/Greenstand/treetracker-admin-client/issues/21)) ([c2933c1](https://github.com/Greenstand/treetracker-admin-client/commit/c2933c164fbeb1fb9719c24ad8c70f6326498590))
+
+# [1.1.0](https://github.com/Greenstand/treetracker-admin-client/compare/v1.0.2...v1.1.0) (2021-02-23)
+
+
+### Features
+
+* add optimized images ([#20](https://github.com/Greenstand/treetracker-admin-client/issues/20)) ([f9e57ec](https://github.com/Greenstand/treetracker-admin-client/commit/f9e57eccb7cd7181fc48f140644dbab2f5877021))
+
 ## [1.0.2](https://github.com/Greenstand/treetracker-admin-client/compare/v1.0.1...v1.0.2) (2021-02-20)